Adding the heading and menu item "components"
The importance of these files is really tied up in the file structure
and how we create shared components in the future. Moving forward we
have a `docs/components` section which will hold very simple (static)
components to used in all docs sections.

Note, this does not provide for components that may require JavaScript
to work as a single script would be included multiple times if inlined
with a component that is used more than once on a page.