Adding the Authipay Side menu
Adding the Authipay side menu. This "component" exists under the
authipay/docs domain. This is here instead of being under components
because Liquid tempaltes (jekylls templating langauge) does not support
yielding content into a `include`. As a result the side menus links need
to be specified statically.

Luckily we can use components here and kind of get around this issue.