Adding the language attr to code blocks incase we need to specify
We would only need to do this if highlightjs is not picking up the
language type well enough. For instance we want all urls to be a single
color but don't want to mark them as strings. Moreover highlightjs
sometimes marks a URL as javascript and another as ruby which won't
work.